♻️ [2025-07-30] 重构代码: 拆分 protocol 把 tcp client 拆出来并修复测试文件和关联文件
